requirements.txt
Werkzeug
Flask
numpy
gevent
h5py
tensorflow
gunicorn
Pillow
matplotlib
